+++ src/app/views/About.svelte
@@ -5,34 +5,16 @@
   import Card from "src/partials/Card.svelte"
   import Heading from "src/partials/Heading.svelte"
   import {router} from "src/app/router"
-  import {people} from "src/engine"
-
-  document.title = "About"
+  import {loadPubkeys, env} from "src/engine"
 
   const hash = import.meta.env.VITE_BUILD_HASH
   const nprofile =
     "nprofile1qqsf03c2gsmx5ef4c9zmxvlew04gdh7u94afnknp33qvv3c94kvwxgspz4mhxue69uhhyetvv9ujuerpd46hxtnfduhsz9rhwden5te0wfjkcctev93xcefwdaexwtcpzdmhxue69uhhqatjwpkx2urpvuhx2ue0vamm57"
   const hodlbodPubkey = "97c70a44366a6535c145b333f973ea86dfdc2d7a99da618c40c64705ad98e322"
-  const coraclePubkey = "8ec86ac9e10979998652068ee6b00223b8e3265aabb3fe28fb6b3b6e294adc96"
 
-  // Provide complete details in case they haven't loaded coracle's profile
-  people.key(coraclePubkey).update($person => ({
-    ...$person,
-    pubkey: coraclePubkey,
-    profile: {
-      ...$person?.profile,
-      name: "Coracle",
-    },
-    zapper: {
-      lnurl:
-        "lnurl1dp68gurn8ghj7em909ek2u3wve6kuep09emk2mrv944kummhdchkcmn4wfk8qtmrdaexzcmvv5tqwy7g",
-      callback: "https://api.geyser.fund/.well-known/lnurlp/coracle",
-      nostrPubkey: "b6dcdddf86675287d1a4e8620d92aa905c258d850bf8cc923d39df1edfee5ee7",
-      allowsNostr: true,
-      maxSendable: 5000000000,
-      minSendable: 1000,
-    },
-  }))
+  loadPubkeys([$env.PLATFORM_PUBKEY])
+
+  document.title = "About"
 </script>
 
 <FlexColumn class="gap-8">
@@ -51,11 +33,7 @@
           All funds donated will be used to support server costs and development.
         </p>
         <div class="flex justify-center">
-          <Anchor
-            modal
-            button
-            accent
-            href={router.at("people").of(coraclePubkey).at("zap").toString()}>
+          <Anchor modal button accent href={router.at("zap").qp({splits: []}).toString()}>
             Zap the developer
           </Anchor>
         </div>