+++ src/partials/Textarea.svelte
@@ -8,7 +8,7 @@
 
   const className = cx(
     $$props.class,
-    "rounded-xl shadow-inset py-2 px-4 pr-10 w-full bg-white text-black",
+    "rounded shadow-inset py-2 px-4 pr-10 w-full bg-white text-black",
     "placeholder:text-mid",
   )
 </script>